Fix numpy float

+++ classes/pyZTD.py
@@ -199,7 +199,7 @@ class Ztd(object):
                 # (missing jump, etc) you end up with very unstable inversions
                 # f[f > 500] = 500
                 sw = np.power(10, LIMIT - s[s > LIMIT])
-                sw[sw < np.finfo(np.float).eps] = np.finfo(np.float).eps
+                sw[sw < np.finfo(float).eps] = np.finfo(float).eps
                 f[s > LIMIT] = sw
 
                 P = np.square(np.divide(f, factor))
@@ -209,7 +209,7 @@ class Ztd(object):
             iteration += 1
 
         # make sure there are no values below eps. Otherwise matrix becomes singular
-        P[P < np.finfo(np.float).eps] = 1e-6
+        P[P < np.finfo(float).eps] = 1e-6
 
         # some statistics
         SS = np.linalg.inv(np.dot(A.transpose(), np.multiply(P[:, None], A)))
